feat(binary): introduced `createStreamDecompress` function
Create a `Transform` stream that decompresses data if it is compressed. If the data is not compressed, it is passed through. This functions allows, stream to be lazily decompress, however, it comes with limitations.

For auto-detection, this function uses Magic headers. This can cause issues if the data is not compressed, but happens to match a Magic header.