feat(functions): added `createIndex` function
This commit adds the `createIndex` function. It creates a function
that returns the next number in a sequence. Once the sequence
reaches the given maximum value, it will start over from the
initial value. By default, the maximum value is the maximum safe
integer and the initial value is zero.