feat(string)!: Changed parameter order in `patternMatch` and `patternMatches` functions
These changes reverse the parameter order for both the `patternMatch` and `patternMatches` functions in the string package. In addition, both functions now use the `createPatternRegexp` function instead of `patternRegex`. This change conforms with many functional programming patterns where the data to be operated on is provided last. This is a breaking change for any implementation that uses these functions.