feat!(string): complete overhaul of `parseJson` function
The `parseJson` function has been completely rewritten. The new implementation allows for more efficient parsing by checking for known JSON values and by returning early. It also checks if the input is a valid JSON string and returns `undefined` if it's not. Furthermore, it throws an error if there's an attempt to parse a JSON string that contains the `__proto__` or `constructor` property, preventing possible prototype pollution attacks. Extensive test cases have been added.