feat(string): added `truncate` function
A new utility function, `truncate`, has been added to the `string` package. This function takes a string and a maximum length and returns the string truncated to the specified length. If the string is longer than the specified length, the last characters are replaced with an ellipsis ('...'). The default ellipsis can be replaced by a custom one. Unit tests have been added to ensure correct functionality.