feat(string!): enhanced `split` functionality and added tests
The `split` function now has default arguments for the delimiter and limit parameters. The default behaviour has been changed to filter out empty strings. The delimiter now defaults to an empty string. The limit can be negative to limit the number of elements from the end of the string. New tests have been added to accommodate these changes.