feat(string): add `entropyShannon` function to compute string entropy
Moved the existing `entropy` function to `entropyShannon` in the `string` package. This function computes the Shannon entropy of a string, which is a measure of its randomness or information it contains. This can be useful in estimating password strength. It's been accompanied by a series of unit tests to verify its functionality.