feat(math)!: changed `Complex` from interface to class
The `Complex` type was previously an interface used to operate with complex numbers. We had to import special functions (add, subtract, multiply, divide, toString) from other modules to operate on them which was not efficient.

Now, it has been converted into a class with methods that allows complex numbers to directly use these operations. This is a breaking change as it changes the way complex numbers are handled in the math package.