fix(string): ensure global replacement in wildcard pattern in `createPatternRegexp`
This change fixes an issue in the `createPatternRegexp` function where only the first instance of a comma in a wildcard pattern within braces `{}` was being replaced. With this fix, all commas within a wildcard pattern are replaced, allowing RegExp to be created correctly for multiple wildcards. In addition, a new test case has been added to test this fix.