feat(types): introduce `MaybeLiteral` and remove `StringNotLiteral`
Introduced a new `MaybeLiteral` type that wraps a literal type in a union with `string` to allow for comprehensive autocompletion and type checking. This new type replaces the now-deleted `StringNotLiteral` type.

Tests were implemented to ensure `MaybeLiteral` behaves as expected. The `MaybeLiteral` type improves the usage of nested properties of an object and can be matched with any string type.