feat(types): new type `NotStringEmpty`
This adds a new TypeScript utility type `NotStringEmpty` to the types package. This type is a union that prevents the user from using an empty string. Test cases have been provided to ensure the correct functioning of the new type. This new type is useful for creating type constraints to prevent use of empty strings.