chore(eslint-config): rename 'dist' directory to 'lib'
Renamed the 'dist' directory to 'lib' in the eslint-config package, updating the "module", "exports", and "files" fields in package.json to reflect this change. This update is in line with the typical convention of using 'lib' as the folder name for distribution files. The build scripts referring to 'dist' have been removed as they are no longer relevant.