feat(collection): refactor 'every' and 'some' functions
Refactored the 'every' and 'some' functions in the collection package. The update provides an improved interface for `every` and `some` function which takes either a function or a value as second argument. It enhances the function to work with both collection and value. Also, upgrades the vitest unit tests for these functions to cover more cases.