chore: enforce the use of `pnpm` via a pre-install script
A pre-install script (`forcePnpm.js`) was added to enforce the use of `pnpm` as the package manager in this project. If another package manager is detected during the install phase, the script will throw an error and halt the install process. The `"preinstall"` command was also added to the `package.json` to trigger this script before every install.