feat(types): improve Get type functionality and tests
Extended the functionality of the `Get` type to support arrays. Now, if the object in question is an array, it will return the item type of the array. Additionally, updated the relevant unit tests to reflect these changes and added a new test case for nested tuples within an object.