feat(collection): add `alias` function
This commit adds a new `alias` function to the `collection` package. This function wraps a collection in a `Proxy` that maps nested properties to top-level for easier access. New utility types `Alias` and `AliasMap` were also introduced in the `types` package to support the `alias` function. Tests were added to validate these new utilities.