docs(eslint-config): enforce consistent Array types
Added a rule to enforce the use of `Array<T>` over `T[]`, and `ReadonlyArray<T>` over `readonly T[]` in the TypeScript ESLint rules. This will help reduce cognitive load when reading code that uses arrays. Also provided link to the rules on the official TypeScript ESLint website for easy reference.