chore(eslint-config): refine import rules and formatting in ESLint config
This commit refines the 'import' rules in the ESLint configuration to impose a consistent ordering for `require()`/`import` statements, thereby reducing unnecessary merge conflicts when multiple people are working on the same file. It also adjusts the formatting of the configuration file for better readability. In the process, deprecated configurations have also been removed, such as 'sort-imports', which is superseded by 'import/order'.