feat(eslint-config): enforce ordering of type union members
The `@typescript-eslint/sort-type-constituents` rule has been added to the ESLint rules. It's set to 'error' to enforce a rule that requires alphabetical ordering of type union members. The intention behind this rule is to reduce cognitive load when dealing with code involving union types. Link to the rules has been provided in the ESLint commenting for easy reference. Some unneeded settings in the `@typescript-eslint/naming-convention` rule are also removed in this commit.