feat(process): add workerRegister and workerRequest
Added `workerRegister` method that allows you to register a function. This can be then called using `node:worker_threads` messaging events.
It allows to organize the flow of the application in a more modular way.

Simultaneously added `workerRequest` which acts as a payload for the IPC message. It is received from the port specified in the request.