feat(process): implement WorkerService for dynamic threading
Added a `WorkerService` class into the process package, that enables execution of functions in a separate worker thread. It has the ability to import and execute functions dynamically anywhere in the application, simplifying the process of implementing worker threads.

In addition, a `Workerized` type and a `WorkerServiceOptions` interface have been created to supplement this feature. The former proxying a module to execute its functions in a worker thread, and the latter including configurations for the service.