feat(process): add WorkerPool class and createWorkerPool function
Introduced the `WorkerPool` class in the `process` package, which represents a pool of workers for executing tasks in parallel. The class has methods for managing workers, executing tasks and terminating the pool. Furthermore, added the `createWorkerPool` function that creates and returns an instance of the `WorkerPool` class. This feature allows tasks to be executed concurrently, improving performance.