chore: added linting for commit messages
Husky and commitlint have been installed and setup for linting of commit messages. The setup includes a new file `.husky/commit-msg` and the `"commit:lint": "commitlint --edit"` script in `package.json`. Commitlint is configured to extend the rules from the `@commitlint/config-pnpm-scopes` config and to allow unlimited length for commit message bodies. All of this is done to ensure that all commit messages follow the same standard.