feat(fs): update `resolveImport` function to not throw exception
The `resolveImport` function in the 'fs' package has been updated and will no longer throw an exception if an import is not found. Instead, it will return `undefined`. This makes the function safer to use and reduces unnecessary logging of errors in the console. Improve the logic of import resolution and include edge cases in the unit tests.