fix(fs): replace while loop with do-while in `findAncestor` function
This commit improves the `findAncestor` function in the 'fs' package. Initially, it was using a while-do loop which was missing files at the root directory. By changing it to a do-while loop, it now correctly returns files present in the root directory. The 'from' parameter now defaults to the current working directory `cwd()` and the function now returns `undefined` instead of throwing an error if no ancestor file is found. Tests have been updated accordingly.