feat(fs): add `updateFile` function to fs package
A new function `updateFile` has been added to the 'fs' package. The function opens a file, updates its content using the provided callback, and then closes it. The function will throw an error if the file does not exist. This new utility enhances the ability to deal with file operations by adding the capacity to update file content in an efficient and concise way. Accompanying tests has also been added.