feat(types): refine `StringLiteral` type to enforce literal strings
This commit refines the `StringLiteral` type to strictly enforce
matching against literal string types only, and not the general
string type. This ensures that the type will resolve to `never` for
broad string types, and will only match specific, literal strings.

Examples in the documentation have been updated to reflect this
change and to provide clear guidance on the expected behavior of
the `StringLiteral` type. Additionally, a redundant test case that
checked non-string types has been removed as it is no longer
relevant to the purpose of `StringLiteral`.