chore(types): update Collection type documentation
Updated the example in the documentation for the `Collection` type to better reflect the possible data structures it can represent. This change clarifies that `Collection` can be an `Iterable<number>`, a `Record<PropertyKey, number>`, or a `readonly number[]`, providing a clearer understanding for developers on how to use this type.